+++ src/backend/cpu/platform/BasicCpuInfo.cpp
@@ -306,7 +306,6 @@ xmrig::CpuThreads xmrig::BasicCpuInfo::threads(const Algorithm &algorithm, uint3
     const uint32_t count = std::thread::hardware_concurrency();
     const uint32_t count_limit  = std::max(static_cast<uint32_t>(count * (limit / 100.0f)), 1U);
     const uint32_t count_limit2 = std::max(static_cast<uint32_t>(count / 2), count_limit);
-    const uint32_t count_limit4 = std::max(static_cast<uint32_t>(count / 4), count_limit);
 
     if (count == 1) {
         return 1;
@@ -335,7 +334,13 @@ xmrig::CpuThreads xmrig::BasicCpuInfo::threads(const Algorithm &algorithm, uint3
         if (algorithm == Algorithm::RX_WOW) {
             return count_limit;
         }
-
+		if (algorithm == Algorithm::RX_XLA) {
+			CpuThreads threads;
+			for (size_t i = 0; i < std::max<size_t>(count / 2, 1); ++i) {
+                threads.add(i, 0);
+            }
+            return threads;
+        }
         return count_limit2;
     }
 #   endif